Move BMI to Self screening module
