+++ lib/src/features/self_screening/bmi/data/model/bmi_log.dart
