+++ lib/src/features/self_screening/bmi/data/model/bmi_nutrition_maping.freezed.dart
