+++ lib/src/features/self_screening/bmi/data/providers/bmi_log_provider.dart
@@ -1,8 +1,8 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:nishauri/src/features/bmi/data/model/bmi_log.dart';
-import 'package:nishauri/src/features/bmi/data/repositories/bmi_log_repository.dart';
-import 'package:nishauri/src/features/bmi/data/services/bmi_log_service.dart';
-import 'package:nishauri/src/features/bmi/presentation/controllers/bmi_log_controller.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/model/bmi_log.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/repositories/bmi_log_repository.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/services/bmi_log_service.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/controllers/bmi_log_controller.dart';
 
 final bmiLogProvider = StateNotifierProvider<BMILogController, AsyncValue<String>>((ref) {
   final service = BMILogService();
