+++ lib/src/features/self_screening/bmi/data/providers/bmi_status_nutrition_provider.dart
@@ -1,6 +1,6 @@
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:nishauri/src/features/bmi/data/repositories/bmi_status_nutrition_repository.dart';
-import 'package:nishauri/src/features/bmi/data/services/bmi_calculator_service.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/repositories/bmi_status_nutrition_repository.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/services/bmi_calculator_service.dart';
 
 final bmiNutritionProvider = FutureProvider((ref)async {
   final repository = BMIStatusNutritionRepository(BMICalculatorService());
