+++ lib/src/features/self_screening/bmi/data/repositories/bmi_log_repository.dart
@@ -1,5 +1,5 @@
-import 'package:nishauri/src/features/bmi/data/model/bmi_log.dart';
-import 'package:nishauri/src/features/bmi/data/services/bmi_log_service.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/model/bmi_log.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/services/bmi_log_service.dart';
 
 class BMILogRepository {
   final BMILogService _service;
