+++ lib/src/features/self_screening/bmi/data/repositories/bmi_status_nutrition_repository.dart
@@ -1,5 +1,5 @@
-import 'package:nishauri/src/features/bmi/data/model/bmi_nutrition_maping.dart';
-import 'package:nishauri/src/features/bmi/data/services/bmi_calculator_service.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/model/bmi_nutrition_maping.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/services/bmi_calculator_service.dart';
 
 class BMIStatusNutritionRepository {
   final BMICalculatorService _service;
