+++ lib/src/features/self_screening/bmi/presentation/pages/BMICalculatorScreen.dart
@@ -4,13 +4,12 @@ import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_svg/flutter_svg.dart';
 import 'package:go_router/go_router.dart';
 import 'package:hooks_riverpod/hooks_riverpod.dart';
-import 'package:nishauri/src/features/auth/data/providers/auth_provider.dart';
 import 'package:nishauri/src/features/auth/data/respositories/auth_repository.dart';
 import 'package:nishauri/src/features/auth/data/services/AuthApiService.dart';
-import 'package:nishauri/src/features/bmi/data/providers/bmi_log_provider.dart';
-import 'package:nishauri/src/features/bmi/presentation/widgets/GenderPicker.dart';
-import 'package:nishauri/src/features/bmi/presentation/widgets/HeightPicker.dart';
-import 'package:nishauri/src/features/bmi/presentation/widgets/HeightUnitsPicker.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/providers/bmi_log_provider.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/widgets/GenderPicker.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/widgets/HeightPicker.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/widgets/HeightUnitsPicker.dart';
 import 'package:nishauri/src/features/user/data/providers/user_provider.dart';
 import 'package:nishauri/src/shared/display/AppCard.dart';
 import 'package:nishauri/src/shared/display/CustomAppBar.dart';
