+++ lib/src/features/self_screening/bmi/presentation/pages/BMIHistoryScreen.dart
@@ -2,9 +2,9 @@ import 'package:fl_chart/fl_chart.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_hooks/flutter_hooks.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:nishauri/src/features/bmi/data/providers/bmi_log_provider.dart';
-import 'package:nishauri/src/features/bmi/presentation/widgets/BMILineGraph.dart';
-import 'package:nishauri/src/features/bmi/presentation/widgets/BMILineList.dart';
+import 'package:nishauri/src/features/self_screening/bmi/data/providers/bmi_log_provider.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/widgets/BMILineGraph.dart';
+import 'package:nishauri/src/features/self_screening/bmi/presentation/widgets/BMILineList.dart';
 import 'package:nishauri/src/shared/display/CustomAppBar.dart';
 import 'package:nishauri/src/shared/display/background_image_widget.dart';
 import 'package:nishauri/src/utils/constants.dart';
